+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Code Ocean capsule - MOSuite - normalize counts
 
+## development version
+
+- Improve the Code Ocean app panel parameter selection and categories. (#3, @phoman14)
+
 ## v3.0
 
 - Use the MOSuite v0.3.0 docker image.